Refactor to separate public api and graph
