+++ famplex/tests/test_graph.py
@@ -20,87 +20,6 @@ def test_in_famplex(famplex_graph, test_input, expected):
     assert famplex_graph.in_famplex(*test_input) == expected
 
 
-@pytest.mark.parametrize('test_input,rel_types,expected',
-                         [(('FPLX', 'Voltage_gated_ion_channels'), None, []),
-                          (('FPLX', 'SCN'), None,
-                           [('FPLX', 'Sodium_channels'),
-                            ('FPLX', 'Voltage_gated_ion_channels')]),
-                          (('FPLX', 'TRP'), None,
-                           [('FPLX', 'Voltage_gated_ion_channels')]),
-                          (('HGNC', 'TRPA1'), None, [('FPLX', 'TRP')]),
-                          (('HGNC', 'PRKAA1'), None,
-                           [('FPLX', 'AMPK_A1B1G1'),
-                            ('FPLX', 'AMPK_A1B1G2'),
-                            ('FPLX', 'AMPK_A1B1G3'),
-                            ('FPLX', 'AMPK_A1B2G1'),
-                            ('FPLX', 'AMPK_A1B2G2'),
-                            ('FPLX', 'AMPK_A1B2G3'),
-                            ('FPLX', 'AMPK_alpha')]),
-                          (('HGNC', 'PRKAA1'), ['isa'],
-                           [('FPLX', 'AMPK_alpha')]),
-                          (('HGNC', 'PRKAA1'), ['partof'],
-                           [('FPLX', 'AMPK_A1B1G1'),
-                            ('FPLX', 'AMPK_A1B1G2'),
-                            ('FPLX', 'AMPK_A1B1G3'),
-                            ('FPLX', 'AMPK_A1B2G1'),
-                            ('FPLX', 'AMPK_A1B2G2'),
-                            ('FPLX', 'AMPK_A1B2G3')])])
-def test_parent_terms(famplex_graph, test_input, rel_types, expected):
-    assert famplex_graph.parent_terms(*test_input,
-                                      relation_types=rel_types) == expected
-
-
-def test_parent_terms_raises(famplex_graph):
-    with pytest.raises(ValueError):
-        famplex_graph.parent_terms('HGNC', 'GENE')
-
-
-@pytest.mark.parametrize('test_input,rel_types,expected',
-                         [(('HGNC', 'TRPA1'), None, []),
-                          (('FPLX', 'ESR'), None,
-                           [('HGNC', 'ESR1'), ('HGNC', 'ESR2')]),
-                          (('FPLX', 'AMPK'), None,
-                           [('FPLX', 'AMPK_A1B1G1'),
-                            ('FPLX', 'AMPK_A1B1G2'),
-                            ('FPLX', 'AMPK_A1B1G3'),
-                            ('FPLX', 'AMPK_A1B2G1'),
-                            ('FPLX', 'AMPK_A1B2G2'),
-                            ('FPLX', 'AMPK_A1B2G3'),
-                            ('FPLX', 'AMPK_A2B1G1'),
-                            ('FPLX', 'AMPK_A2B1G2'),
-                            ('FPLX', 'AMPK_A2B1G3'),
-                            ('FPLX', 'AMPK_A2B2G1'),
-                            ('FPLX', 'AMPK_A2B2G2'),
-                            ('FPLX', 'AMPK_A2B2G3'),
-                            ('FPLX', 'AMPK_alpha'),
-                            ('FPLX', 'AMPK_beta'),
-                            ('FPLX', 'AMPK_gamma')]),
-                          (('FPLX', 'AMPK'), ['isa'],
-                          [('FPLX', 'AMPK_A1B1G1'),
-                           ('FPLX', 'AMPK_A1B1G2'),
-                           ('FPLX', 'AMPK_A1B1G3'),
-                           ('FPLX', 'AMPK_A1B2G1'),
-                           ('FPLX', 'AMPK_A1B2G2'),
-                           ('FPLX', 'AMPK_A1B2G3'),
-                           ('FPLX', 'AMPK_A2B1G1'),
-                           ('FPLX', 'AMPK_A2B1G2'),
-                           ('FPLX', 'AMPK_A2B1G3'),
-                           ('FPLX', 'AMPK_A2B2G1'),
-                           ('FPLX', 'AMPK_A2B2G2'),
-                           ('FPLX', 'AMPK_A2B2G3')]),
-                          (('FPLX', 'AMPK'), ['partof'],
-                          [('FPLX', 'AMPK_alpha'),
-                           ('FPLX', 'AMPK_beta'),
-                           ('FPLX', 'AMPK_gamma')])])
-def test_child_terms(famplex_graph, test_input, rel_types, expected):
-    assert famplex_graph.child_terms(*test_input,
-                                     relation_types=rel_types) == expected
-
-
-def test_child_terms_raises(famplex_graph):
-    with pytest.raises(ValueError):
-        famplex_graph.child_terms('FPLX', 'Complex')
-
 @pytest.mark.parametrize('test_input,expected',
                          [(('HGNC', 'ESR1'), [('FPLX', 'ESR')]),
                           (('FPLX', 'ESR'), [('FPLX', 'ESR')]),
@@ -114,222 +33,3 @@ def test_root_terms(famplex_graph, test_input, expected):
 def test_root_terms_raises(famplex_graph):
     with pytest.raises(ValueError):
         famplex_graph.root_terms('HGNC', 'Gene')
-
-
-@pytest.mark.parametrize('test_input,rel_types,expected',
-                         [(('FPLX', 'ESR'), None, []),
-                          (('HGNC', 'ESR1'), None, [('FPLX', 'ESR')]),
-                          (('HGNC', 'SCN8A'), None,
-                          [('FPLX',
-                            'Sodium_voltage_gated_channel_alpha_subunits'),
-                           ('FPLX', 'SCN'), ('FPLX', 'Sodium_channels'),
-                           ('FPLX', 'Voltage_gated_ion_channels'),
-                           ('FPLX', 'Cation_channels')]),
-                          (('HGNC', 'PRKAA1'), None,
-                           [('FPLX', 'AMPK_A1B1G1'),
-                            ('FPLX', 'AMPK_A1B1G2'),
-                            ('FPLX', 'AMPK_A1B1G3'),
-                            ('FPLX', 'AMPK_A1B2G1'),
-                            ('FPLX', 'AMPK_A1B2G2'),
-                            ('FPLX', 'AMPK_A1B2G3'),
-                            ('FPLX', 'AMPK_alpha'),
-                            ('FPLX', 'AMPK')]),
-                          (('HGNC', 'PRKAA1'), ['isa'],
-                           [('FPLX', 'AMPK_alpha')]),
-                          (('HGNC', 'PRKAA1'), ['partof'],
-                           [('FPLX', 'AMPK_A1B1G1'),
-                            ('FPLX', 'AMPK_A1B1G2'),
-                            ('FPLX', 'AMPK_A1B1G3'),
-                            ('FPLX', 'AMPK_A1B2G1'),
-                            ('FPLX', 'AMPK_A1B2G2'),
-                            ('FPLX', 'AMPK_A1B2G3')])])
-def test_ancestral_terms(famplex_graph, test_input, rel_types, expected):
-    assert famplex_graph.ancestral_terms(*test_input,
-                                         relation_types=rel_types) == expected
-
-
-def test_ancestral_terms_raises(famplex_graph):
-    with pytest.raises(ValueError):
-        famplex_graph.ancestral_terms('HGNC', 'GENE')
-
-
-@pytest.mark.parametrize('test_input,rel_types,expected',
-                         [(('HGNC', 'ESR1'), None, []),
-                          (('FPLX', 'AMPK'), None,
-                           [('FPLX', 'AMPK_A1B1G1'),
-                            ('FPLX', 'AMPK_A1B1G2'),
-                            ('FPLX', 'AMPK_A1B1G3'),
-                            ('FPLX', 'AMPK_A1B2G1'),
-                            ('FPLX', 'AMPK_A1B2G2'),
-                            ('FPLX', 'AMPK_A1B2G3'),
-                            ('FPLX', 'AMPK_A2B1G1'),
-                            ('FPLX', 'AMPK_A2B1G2'),
-                            ('FPLX', 'AMPK_A2B1G3'),
-                            ('FPLX', 'AMPK_A2B2G1'),
-                            ('FPLX', 'AMPK_A2B2G2'),
-                            ('FPLX', 'AMPK_A2B2G3'),
-                            ('FPLX', 'AMPK_alpha'),
-                            ('FPLX', 'AMPK_beta'),
-                            ('FPLX', 'AMPK_gamma'),
-                            ('HGNC', 'PRKAA1'),
-                            ('HGNC', 'PRKAB1'),
-                            ('HGNC', 'PRKAG1'),
-                            ('HGNC', 'PRKAG2'),
-                            ('HGNC', 'PRKAG3'),
-                            ('HGNC', 'PRKAB2'),
-                            ('HGNC', 'PRKAA2')]),
-                          (('FPLX', 'AMPK'), ['isa'],
-                           [('FPLX', 'AMPK_A1B1G1'),
-                            ('FPLX', 'AMPK_A1B1G2'),
-                            ('FPLX', 'AMPK_A1B1G3'),
-                            ('FPLX', 'AMPK_A1B2G1'),
-                            ('FPLX', 'AMPK_A1B2G2'),
-                            ('FPLX', 'AMPK_A1B2G3'),
-                            ('FPLX', 'AMPK_A2B1G1'),
-                            ('FPLX', 'AMPK_A2B1G2'),
-                            ('FPLX', 'AMPK_A2B1G3'),
-                            ('FPLX', 'AMPK_A2B2G1'),
-                            ('FPLX', 'AMPK_A2B2G2'),
-                            ('FPLX', 'AMPK_A2B2G3')]),
-                          (('FPLX', 'AMPK'), ['partof'],
-                           [('FPLX', 'AMPK_alpha'),
-                            ('FPLX', 'AMPK_beta'),
-                            ('FPLX', 'AMPK_gamma')])])
-def test_descendant_terms(famplex_graph, test_input, rel_types, expected):
-    assert famplex_graph.descendant_terms(*test_input,
-                                          relation_types=rel_types) == expected
-
-
-def test_descendant_terms_raises(famplex_graph):
-    with pytest.raises(ValueError):
-        famplex_graph.descendant_terms('FPLX', 'Complex')
-
-
-@pytest.mark.parametrize('test_input,rel_types,expected',
-                         [(('FPLX', 'ESR'), None,
-                           [('HGNC', 'ESR1'), ('HGNC', 'ESR2')]),
-                          (('FPLX', 'AMPK'), None,
-                           [('HGNC', 'PRKAA1'),
-                            ('HGNC', 'PRKAA2'),
-                            ('HGNC', 'PRKAB1'),
-                            ('HGNC', 'PRKAB2'),
-                            ('HGNC', 'PRKAG1'),
-                            ('HGNC', 'PRKAG2'),
-                            ('HGNC', 'PRKAG3')]),
-                          (('FPLX', 'AMPK'), ['isa'],
-                           [('FPLX', 'AMPK_A1B1G1'),
-                            ('FPLX', 'AMPK_A1B1G2'),
-                            ('FPLX', 'AMPK_A1B1G3'),
-                            ('FPLX', 'AMPK_A1B2G1'),
-                            ('FPLX', 'AMPK_A1B2G2'),
-                            ('FPLX', 'AMPK_A1B2G3'),
-                            ('FPLX', 'AMPK_A2B1G1'),
-                            ('FPLX', 'AMPK_A2B1G2'),
-                            ('FPLX', 'AMPK_A2B1G3'),
-                            ('FPLX', 'AMPK_A2B2G1'),
-                            ('FPLX', 'AMPK_A2B2G2'),
-                            ('FPLX', 'AMPK_A2B2G3')]),
-                          (('FPLX', 'AMPK'), ['partof'],
-                           [('FPLX', 'AMPK_alpha'),
-                            ('FPLX', 'AMPK_beta'),
-                            ('FPLX', 'AMPK_gamma')])])
-def test_individual_members(famplex_graph, test_input, rel_types, expected):
-    assert famplex_graph.\
-        individual_members(*test_input,
-                           relation_types=rel_types) == expected
-
-
-def test_individual_members_raises(famplex_graph):
-    with pytest.raises(ValueError):
-        famplex_graph.individual_members('FPLX', 'Complex')
-
-
-@pytest.mark.parametrize('test_input,expected',
-                         [(('HGNC', 'ESR1', 'FPLX', 'ESR'), True),
-                          (('FPLX', 'ESR', 'HGNC', 'ESR1'), False),
-                          (('HGNC', 'SCN8A', 'FPLX', 'SCN'), True),
-                          (('HGNC', 'PRKAB1', 'FPLX', 'AMPK_A2B1G1'), False),
-                          (('FPLX', 'AMPK_A2B1G1', 'HGNC', 'PRKAB1'), False),
-                          (('FPLX', 'AMPK_A2B1G1', 'FPLX', 'AMPK'), True),
-                          (('HGNC', 'PRKAA1', 'FPLX', 'AMPK'), False),
-                          (('HGNC', 'DAP3', 'FPLX',
-                            'Mitochondrial_Ribosome'), False),
-                          (('HGNC', 'SCN8A', 'FPLX', 'MEK'), False),
-                          (('HGNC', 'GENE', 'FPLX', 'MEK'), False)])
-def test_isa(famplex_graph, test_input, expected):
-    assert famplex_graph.isa(*test_input) == expected
-
-
-@pytest.mark.parametrize('test_input,expected',
-                         [(('HGNC', 'ESR1', 'FPLX', 'ESR'), False),
-                          (('FPLX', 'ESR', 'HGNC', 'ESR1'), False),
-                          (('HGNC', 'SCN8A', 'FPLX', 'SCN'), False),
-                          (('HGNC', 'PRKAB1', 'FPLX', 'AMPK_A2B1G1'), True),
-                          (('FPLX', 'AMPK_A2B1G1', 'HGNC', 'PRKAB1'), False),
-                          (('FPLX', 'AMPK_A2B1G1', 'FPLX', 'AMPK'), False),
-                          (('HGNC', 'PRKAA1', 'FPLX', 'AMPK'), False),
-                          (('HGNC', 'DAP3', 'FPLX',
-                            'Mitochondrial_Ribosome'), True),
-                          (('HGNC', 'SCN8A', 'FPLX', 'MEK'), False),
-                          (('HGNC', 'GENE', 'FPLX', 'MEK'), False)])
-def test_partof(famplex_graph, test_input, expected):
-    assert famplex_graph.partof(*test_input) == expected
-
-
-@pytest.mark.parametrize('test_input,expected',
-                         [(('HGNC', 'ESR1', 'FPLX', 'ESR'), True),
-                          (('FPLX', 'ESR', 'HGNC', 'ESR1'), False),
-                          (('HGNC', 'SCN8A', 'FPLX', 'SCN'), True),
-                          (('HGNC', 'PRKAB1', 'FPLX', 'AMPK_A2B1G1'), True),
-                          (('FPLX', 'AMPK_A2B1G1', 'HGNC', 'PRKAB1'), False),
-                          (('FPLX', 'AMPK_A2B1G1', 'FPLX', 'AMPK'), True),
-                          (('HGNC', 'PRKAA1', 'FPLX', 'AMPK'), True),
-                          (('HGNC', 'DAP3', 'FPLX',
-                            'Mitochondrial_Ribosome'), True),
-                          (('HGNC', 'SCN8A', 'FPLX', 'MEK'), False),
-                          (('HGNC', 'GENE', 'FPLX', 'MEK'), False)])
-def test_refinement_of(famplex_graph, test_input, expected):
-    assert famplex_graph.refinement_of(*test_input) == expected
-
-
-@pytest.mark.parametrize('test_input,expected',
-                         # Estrogen Receptor Family
-                         [(('FPLX', 'ESR'),
-                           # expected
-                           {('FPLX', 'ESR'):
-                            [({('HGNC', 'ESR1'): []}, 'isa'),
-                             ({('HGNC', 'ESR2'): []}, 'isa')]}),
-                          # Estrogen Receptor 1
-                          (('HGNC', 'ESR1'),
-                           # expected
-                           {('HGNC', 'ESR1'): []}),
-                          # MAP2K
-                          (('FPLX', 'MAP2K'),
-                           # expected
-                           {('FPLX', 'MAP2K'):
-                            [({('FPLX', 'MEK'):
-                               [({('HGNC', 'MAP2K1'): []}, 'isa'),
-                                ({('HGNC', 'MAP2K2'): []}, 'isa')]},
-                              'isa'),
-                             ({('HGNC', 'MAP2K3'): []}, 'isa'),
-                             ({('HGNC', 'MAP2K4'): []}, 'isa'),
-                             ({('HGNC', 'MAP2K5'): []}, 'isa'),
-                             ({('HGNC', 'MAP2K6'): []}, 'isa'),
-                             ({('HGNC', 'MAP2K7'): []}, 'isa')]})])
-def test_dict_representation(famplex_graph, test_input, expected):
-    assert famplex_graph.dict_representation(*test_input) == expected
-
-
-def test_dict_representation_raises(famplex_graph):
-    with pytest.raises(ValueError):
-        famplex_graph.dict_representation('FPLX', 'Complex')
-
-
-def test_equivalences(famplex_graph):
-    assert set([('BEL', 'ESR Family')]) <= \
-        set(famplex_graph.equivalences('ESR'))
-
-
-def test_equivalences_raises(famplex_graph):
-    with pytest.raises(ValueError):
-        famplex_graph.equivalences('Complex')
